e-commerce/backend/pyproject.toml
[tool.poetry]
name = "ecommerce-backend"
version = "0.1.0"
description = "Multi-tenant FastAPI services for the Premium E-Commerce platform."
authors = ["AI Assistant <assistant@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = { version = "^0.30.3", extras = ["standard"] }
SQLAlchemy = "^2.0.32"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
structlog = "^24.1.0"
tenacity = "^9.0.0"
redis = {extras = ["hiredis"], version = "^5.0.7"}
httpx = "^0.27.0"
orjson = "^3.10.6"
celery = {extras = ["redis"], version = "^5.4.0"}
kombu = "^5.4.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg2-binary = "^2.9.9"
stripe = "^8.0.0"
razorpay = "^1.4.0"
setuptools = "^75.0.0"
pandas = "^2.2.0"
openpyxl = "^3.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
pytest-cov = "^4.1.0"
ruff = "^0.5.7"
mypy = "^1.11.1"
pre-commit = "^3.7.0"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

